+++ effects/$storageKey.ts
@@ -1,5 +1,7 @@
-import * as Z from "../../deps/zones.ts";
-import * as M from "../../frame_metadata/mod.ts";
+import * as Z from "../deps/zones.ts";
+import * as M from "../frame_metadata/mod.ts";
+
+const k0_ = Symbol();
 
 export const $storageKey = Z.call.fac((
   deriveCodec: M.DeriveCodec,
@@ -11,4 +13,4 @@ export const $storageKey = Z.call.fac((
     pallet,
     storageEntry,
   });
-});
+}, k0_);
