+++ effects/blockRead.ts
@@ -1,10 +1,12 @@
 import * as Z from "../deps/zones.ts";
 import * as rpc from "../rpc/mod.ts";
 import * as U from "../util/mod.ts";
-import { $extrinsic } from "./core/$extrinsic.ts";
-import { deriveCodec } from "./core/deriveCodec.ts";
+import { $extrinsic } from "./$extrinsic.ts";
+import { deriveCodec } from "./deriveCodec.ts";
 import { metadata } from "./metadata.ts";
-import { chain } from "./rpc/known.ts";
+import { chain } from "./rpc_known.ts";
+
+const k0_ = Symbol();
 
 export function blockRead<Client extends Z.$<rpc.Client>>(client: Client) {
   return <Rest extends [blockHash?: Z.$<U.HexHash | undefined>]>(...[blockHash]: [...Rest]) => {
@@ -13,7 +15,7 @@ export function blockRead<Client extends Z.$<rpc.Client>>(client: Client) {
     const call = chain.getBlock(client)(blockHash);
     return Z
       .ls($extrinsic_, call)
-      .next(function mapExtrinsicCall([$extrinsic_, call]) {
+      .next(([$extrinsic_, call]) => {
         const { block: { extrinsics, header }, justifications } = call;
         return {
           justifications,
@@ -24,7 +26,7 @@ export function blockRead<Client extends Z.$<rpc.Client>>(client: Client) {
             }),
           },
         };
-      })
+      }, k0_)
       .zoned("BlockRead");
   };
 }
