+++ effects/blockWatch.ts
@@ -4,7 +4,9 @@ import * as known from "../known/mod.ts";
 import * as rpc from "../rpc/mod.ts";
 import * as U from "../util/mod.ts";
 import { blockRead } from "./blockRead.ts";
-import { chain } from "./rpc/known.ts";
+import { chain } from "./rpc_known.ts";
+
+const k0_ = Symbol();
 
 export function blockWatch<Client extends Z.$<rpc.Client>>(client: Client) {
   return <
@@ -12,14 +14,14 @@ export function blockWatch<Client extends Z.$<rpc.Client>>(client: Client) {
   >(listener: Listener) => {
     const listenerMapped = Z
       .ls(listener, Z.env)
-      .next(function mapBlockWatchListener([listener, env]) {
+      .next(([listener, env]) => {
         return async function(this: rpc.ClientSubscribeContext, header: known.Header) {
           const blockHash = chain.getBlockHash(client)(header.number);
           const block = await blockRead(client)(blockHash).bind(env)();
           if (block instanceof Error) throw block;
           listener.apply(this, [block]);
         };
-      });
+      }, k0_);
     const subscriptionId = chain.subscribeNewHeads(client)([], listenerMapped);
     return chain
       .unsubscribeNewHeads(client)(subscriptionId)
