+++ effects/const.ts
@@ -1,10 +1,10 @@
 import * as Z from "../deps/zones.ts";
 import * as rpc from "../rpc/mod.ts";
 import * as U from "../util/mod.ts";
-import { codec } from "./core/codec.ts";
-import { deriveCodec } from "./core/deriveCodec.ts";
-import * as e$ from "./core/scale.ts";
+import { codec } from "./codec.ts";
+import { deriveCodec } from "./deriveCodec.ts";
 import { constMetadata, metadata, palletMetadata } from "./metadata.ts";
+import * as e$ from "./scale.ts";
 
 export function const_<Client extends Z.$<rpc.Client>>(client: Client) {
   return <
@@ -23,7 +23,7 @@ export function const_<Client extends Z.$<rpc.Client>>(client: Client) {
     const entryValueTypeI = constMetadata_.access("ty").access("id");
     const constValue = constMetadata_.access("value");
     const $const = codec(deriveCodec_, entryValueTypeI);
-    return e$.decoded($const, constValue, "value").zoned("Const");
+    return e$.scaleDecoded($const, constValue, "value").zoned("Const");
   };
 }
 Object.defineProperty(const_, "name", {
