+++ effects/metadata.ts
@@ -3,20 +3,26 @@ import * as Z from "../deps/zones.ts";
 import * as M from "../frame_metadata/mod.ts";
 import * as rpc from "../rpc/mod.ts";
 import * as U from "../util/mod.ts";
-import { state } from "./rpc/known.ts";
+import { state } from "./rpc_known.ts";
+
+const k0_ = Symbol();
+const k1_ = Symbol();
+const k2_ = Symbol();
+const k3_ = Symbol();
+const k4_ = Symbol();
 
 // TODO: callable object so that one doesn't need the extra parens when not specifying block hash?
 export function metadata<Client extends Z.$<rpc.Client>>(client: Client) {
   return <Rest extends [blockHash?: Z.$<U.HexHash | undefined>]>(...[blockHash]: [...Rest]) => {
     return state
       .getMetadata(client)(blockHash)
-      .next(function metadataImpl(encoded) {
+      .next((encoded) => {
         try {
           return M.fromPrefixedHex(encoded);
         } catch (e) {
           return e as $.ScaleError;
         }
-      })
+      }, k0_)
       .zoned("Metadata");
   };
 }
@@ -27,9 +33,9 @@ export function palletMetadata<Metadata extends Z.$<M.Metadata>, PalletName exte
 ) {
   return Z
     .ls(metadata, palletName)
-    .next(function palletMetadataImpl([metadata, palletName]) {
+    .next(([metadata, palletName]) => {
       return M.getPallet(metadata, palletName);
-    })
+    }, k1_)
     .zoned("PalletMetadata");
 }
 
@@ -39,9 +45,9 @@ export function entryMetadata<PalletMetadata extends Z.$<M.Pallet>, EntryName ex
 ) {
   return Z
     .ls(palletMetadata, entryName)
-    .next(function entryMetadataImpl([palletMetadata, entryName]) {
+    .next(([palletMetadata, entryName]) => {
       return M.getEntry(palletMetadata, entryName);
-    })
+    }, k2_)
     .zoned("EntryMetadata");
 }
 
@@ -54,9 +60,9 @@ export function constMetadata<
 ) {
   return Z
     .ls(palletMetadata, constName)
-    .next(function constMetadataImpl([palletMetadata, constName]) {
+    .next(([palletMetadata, constName]) => {
       return M.getConst(palletMetadata, constName);
-    })
+    }, k3_)
     .zoned("ConstMetadata");
 }
 
@@ -66,14 +72,14 @@ export function mapMetadata<PalletMetadata extends Z.$<M.Pallet>, EntryName exte
 ) {
   return Z
     .ls(palletMetadata, entryName)
-    .next(function mapMetadataImpl([palletMetadata, entryName]) {
+    .next(([palletMetadata, entryName]) => {
       const entryMetadata = M.getEntry(palletMetadata, entryName);
       if (entryMetadata instanceof Error) return entryMetadata;
       if (entryMetadata.type !== "Map") {
         return new ExpectedMapError();
       }
       return entryMetadata;
-    })
+    }, k4_)
     .zoned("MapMetadata");
 }
