+++ test_util/extrinsic.ts
@@ -4,14 +4,14 @@ import { ExtrinsicProps, SignedExtrinsic } from "../effects/extrinsic.ts";
 import * as M from "../frame_metadata/mod.ts";
 import * as known from "../known/rpc/mod.ts";
 
+const k0_ = Symbol();
+
 // TODO: use context / properly scope context / make it accessible outside of subscription lifecycle
 // TODO: better zones-level way to share context between effects
 export function collectExtrinsicEvents<
   Props extends Z.Rec$<ExtrinsicProps>,
   Sign extends Z.$<M.SignExtrinsic>,
->(
-  extrinsic: SignedExtrinsic<Props, Sign>,
-) {
+>(extrinsic: SignedExtrinsic<Props, Sign>) {
   const events: known.TransactionStatus[] = [];
   return extrinsic
     .watch(function(status) {
@@ -20,7 +20,7 @@ export function collectExtrinsicEvents<
         this.stop();
       }
     })
-    .next(() => events);
+    .next(() => events, k0_);
 }
 
 // TODO: is this a common-enough test to merit existence in `test_util`?
